% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Generate synthetic data for survival analysis}
\usage{
data_generation(n, N, lam)
}
\arguments{
\item{n}{Integer specifying the number of labeled data points to generate}

\item{N}{Integer specifying the number of unlabeled data points to generate}

\item{lam}{Numeric value between 0 and 1 representing the proportion of event and censoring times used to generate the covariates}
}
\value{
A list containing the following generated data:
- `Ti`: Event times for the labeled data
- `Di`: Event indicators for the labeled data (1 for event, 0 for censored)
- `Ci`: Censoring times for the labeled data
- `Zi`: Covariate values for the labeled data
- `Xi`: Minimum of event and censoring times for the labeled data
- `Ti.UL`: Event times for the unlabeled data
- `Ci.UL`: Censoring times for the unlabeled data
- `Zi.UL`: Covariate values for the unlabeled data
- `Xi.UL`: Minimum of event and censoring times for the unlabeled data
}
\description{
This function generates synthetic labeled and unlabeled data for survival analysis, simulating event times, censoring times, and covariates.
}
