% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.DL.R
\name{get.DL}
\alias{get.DL}
\title{Calculate Deep Learning-based estimator}
\usage{
get.DL(Zi, Zi.UL, Xi, Ci, Di, t0.all)
}
\arguments{
\item{Zi}{A matrix of covariates for the labeled data}

\item{Zi.UL}{A matrix of covariates for the unlabeled data}

\item{Xi}{A numeric vector representing event/censoring times for the labeled data}

\item{Ci}{A numeric vector representing censoring times for the labeled data}

\item{Di}{A numeric vector representing event indicators for the labeled data}

\item{t0.all}{A numeric vector representing the time points at which to estimate survival probabilities}
}
\value{
A data frame containing the time points and the corresponding deep learning-based estimates of survival probabilities
}
\description{
This function computes the deep learning-based estimator for survival probabilities
at each specified time point using a deep survival model.
}
