% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.DR.R
\name{get.DR}
\alias{get.DR}
\title{Compute Density Ratio estimator of survival function}
\usage{
get.DR(Zi, Zi.UL, Ti, Ci, Di, t0.all)
}
\arguments{
\item{Zi}{Matrix of covariate values for the labeled data}

\item{Zi.UL}{Matrix of covariate values for the unlabeled data}

\item{Ti}{Vector of event times for the labeled data}

\item{Ci}{Vector of censoring times for the labeled data}

\item{Di}{Vector of the event indicators for the labeled data (1 for event, 0 for censored)}

\item{t0.all}{Vector of time points at which the survival function is to be estimated}
}
\value{
A list containing the Density Ratio estimator of the survival function (`DR`) at the specified time points in `t0.all`
}
\description{
This function calculates the Density Ratio (DR) estimator of the survival function at the time points specified in `t0.all`.
}
