% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.KM.R
\name{get.KM}
\alias{get.KM}
\title{Semi-parametric estimator with influence function}
\usage{
get.KM(Xi, Di, t0.all)
}
\arguments{
\item{Xi}{Vector of minimum of T and C for labeled data}

\item{Di}{Vector of event indicators for labeled data}

\item{t0.all}{Vector of time points at which the survival functions are estimated}
}
\value{
A list containing the supervised KM estimator and the influence function
}
\description{
This function calculates the supervised KM estimator for survival data.
}
