% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.NP.R
\name{get.NP}
\alias{get.NP}
\title{Non-parametric estimator}
\usage{
get.NP(Xi, Ci, Di, Zi, Ci.UL, Zi.UL, h = 0.2)
}
\arguments{
\item{Xi}{Vector of minimum of T and C for labeled data}

\item{Ci}{Vector of censoring times for labeled data}

\item{Di}{Vector of event indicators for labeled data}

\item{Zi}{Vector of covariates for labeled data}

\item{Ci.UL}{Vector of censoring times for unlabeled data}

\item{Zi.UL}{Vector of covariates for unlabeled data}

\item{h}{Bandwidth parameter for the kernel density estimation (default is set to 0.2)}
}
\value{
A list containing the non-parametric estimator
}
\description{
This function calculates the non-parametric estimator for survival data
}
