% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.semi.auc.R
\name{get.semi.auc}
\alias{get.semi.auc}
\title{Calculate AUC for the semi-parametric model}
\usage{
get.semi.auc(Xi, Ci, Di, Zi, Ci.UL, Zi.UL, t0.all)
}
\arguments{
\item{Xi}{A vector of event times for the labeled data.}

\item{Ci}{A vector of censoring times for the labeled data.}

\item{Di}{A vector of event indicators for the labeled data.}

\item{Zi}{A matrix of covariates for the labeled data.}

\item{Ci.UL}{A vector of censoring times for the unlabeled data.}

\item{Zi.UL}{A matrix of covariates for the unlabeled data.}

\item{t0.all}{A vector of time points at which to calculate the AUC.}
}
\value{
A vector containing the mean AUC for each time point in t0.all.
}
\description{
This function calculates the area under the curve (AUC) for the
semi-parametric model using cross-validation.
}
