% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.semi.R
\name{get.semi}
\alias{get.semi}
\title{Semi-supervised estimator with influence function}
\usage{
get.semi(Xi, Ci, Di, Zi, Ci.UL, Zi.UL, t0.all)
}
\arguments{
\item{Xi}{Vector of minimum of T and C for labeled data}

\item{Ci}{Vector of censoring times for labeled data}

\item{Di}{Vector of event indicators for labeled data}

\item{Zi}{Vector of covariates for labeled data}

\item{Ci.UL}{Vector of censoring times for unlabeled data}

\item{Zi.UL}{Vector of covariates for unlabeled data}
}
\value{
A list containing the semi-parametric estimator and the influence function
}
\description{
This function calculates the semi-supervised estimator for survival data.
}
