% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_DR.R
\name{Newton_DR}
\alias{Newton_DR}
\title{Newton-DR Estimation for  Survival Data}
\usage{
Newton_DR(
  phi.t,
  phi.v,
  weights = 1,
  max.iter = 100,
  tol = 1e-05,
  initial = rep(0, ncol(phi.t)),
  lambda0 = NULL
)
}
\arguments{
\item{phi.t}{A matrix with labeled data, where rows are observations and columns are covariates.}

\item{phi.v}{A matrix with unlabeled data, where rows are observations and columns are covariates.}

\item{weights}{An optional vector of weights for each observation. Default is 1 for equal weights.}

\item{max.iter}{Maximum number of iterations for the Newton-Raphson algorithm. Default is 100.}

\item{tol}{Tolerance for convergence of the Newton-Raphson algorithm. Default is 1e-5.}

\item{initial}{An optional vector of initial values for the algorithm. Default is a vector of zeros with length equal to the number of columns in phi.t.}

\item{lambda0}{An optional initial value for the regularization parameter. If NULL, lambda0 is set to log(ncol(phi.t))/n.t^1.5. Default is NULL.}
}
\value{
A vector of parameter estimates for the transformation model.
}
\description{
The `Newton_DR` function is an iterative algorithm to estimate the parameters in a
transformation model for survival data.
}
