% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rC.fun}
\alias{rC.fun}
\title{Generate random censoring times}
\usage{
rC.fun(nn)
}
\arguments{
\item{nn}{Integer specifying the number of random censoring times to generate}
}
\value{
A vector of nn random censoring times
}
\description{
This function generates random censoring times using an exponential distribution.
}
