% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rT.fun}
\alias{rT.fun}
\title{Generate random event times}
\usage{
rT.fun(nn)
}
\arguments{
\item{nn}{Integer specifying the number of random event times to generate}
}
\value{
A vector of nn random event times
}
\description{
This function generates random event times using an exponential distribution.
}
