\name{SCANER-package}
\alias{SCANER-package}
\title{single cell CNA analysis toolkits}
\description{
single cell CNA analysis toolkits
}
\usage{
CpR to ...
identificatioinTumorCells to identify tumor cells in bulk tissue
}
\arguments{
  \item{name}{a word}
}
\details{
  nothing
}
\value{
  no return
}
\references{
  nothing
}
