% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VTM}
\alias{VTM}
\title{Create a matrix with identical rows}
\usage{
VTM(vc, dm)
}
\arguments{
\item{vc}{Numeric vector to be used as the template for each row of the matrix}

\item{dm}{Integer specifying the number of rows in the output matrix}
}
\value{
An dm x n matrix with each row equal to vc
}
\description{
This function creates a matrix with dm rows and n columns, where n is the length of the input vector vc. Each row of the matrix is identical to vc.
}
\examples{
# Example usage of VTM function
vec <- c(1, 2, 3)
mat <- VTM(vec, 4)
}
